6/12/2025 upload
fixed the non-detections on the lightcurve plots, same with the brightness temperature plots. Still working on the spectral index. Also all new data for V1535 Sco, so the new data are in the notebooks and it is fixed. Also adding in the three new data files (through excell) for the corrected V1535 Sco, and new data for V1534Sco and Rs Oph.