Uploading spectral index work 6/12/2025
Having issues with the spectral index, the code is right, but I am getting the wrong param_cov from the fit_info. Bella was able to get the correct param_cov, working on it to figure out how to get the right param_cov through my laptop. Also updating the V1535Sco work because of incorrect datafile being used originally, this is the one with the correct data.